pyproject.toml
[project]
name = "fastapi-project"
version = "1.0.0"
description = "A FastAPI project with PostgreSQL database and LangChain integration"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "sqlalchemy",
    "psycopg2-binary",
    "alembic",
    "langchain",
    "langchain-core",
    "langchain-community",
    "langchain-openai",
    "openai",
    "tiktoken"
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 